Update README.md
Initially, convert the pdf files into csv files by using fitz. Then upload the files in a database using MySQL workbench. Then create a frontend with the help of HTML and CSS. To create an efficiently working website, connect the frontend with the backend by using flask. In the flask code, add the SQL queries to get the desired result. Hence, the website is ready.